// src/features/dashboard/types/metric-type.ts
export interface MetricType {
  totalTrades: number;
  totalTradesPastWeek: number;
  totalTradesCurrentMonth: number;
  winRate: number;
  profitFactor: number;
  riskRewardRatio: number;
  mostTradedPair: string;
  totalProfit: number;
  totalLoss: number;
  totalNetProfit: number;
  profitLossRatio: number;
  avgProfitPerTrade: number;
}